+++ app/api/statistics/route.ts
@@ -1,0 +1,10 @@
+import { fetchWebsiteStats } from "@/app/lib/data";
+import { NextResponse } from "next/server";
+
+export async function GET() {
+	const stats = await fetchWebsiteStats()
+	console.log(stats)
+	return new NextResponse(JSON.stringify(stats), {
+		headers: { 'Content-Type': 'application/json' },
+	})
+}